Renames BackupRecoveryPhraseScreen to BackupSeedPhraseScreen for better clarity, it was also too similar to the new BackupRestore screen's name; Adds a loading prop to SeedPhraseTable component, to avoid showing an empty state where it shouldn't
