+++ cardstack/src/hooks/useSelectedWallet.ts
@@ -26,6 +26,8 @@ export const useSelectedWallet = () => {
     if (walletReady && !seedPhrase) {
       getSeedPhrase();
     }
+
+    console.log(':::', { walletReady, seedPhrase });
   }, [walletReady, getSeedPhrase, seedPhrase]);
 
   return {
