+++ cardstack/src/screens/Backup/BackupSeedPhraseScreen/strings.ts
