+++ cardstack/src/screens/Backup/index.ts
@@ -1,6 +1,6 @@
 export { default as BackupExplanationScreen } from './BackupExplanationScreen/BackupExplanationScreen';
 export { default as BackupCloudPasswordScreen } from './BackupCloudPasswordScreen/BackupCloudPasswordScreen';
-export { default as BackupRecoveryPhraseScreen } from './BackupRecoveryPhraseScreen/BackupRecoveryPhraseScreen';
+export { default as BackupSeedPhraseScreen } from './BackupSeedPhraseScreen/BackupSeedPhraseScreen';
 export { default as BackupSeedPhraseConfirmationScreen } from './BackupSeedPhraseConfirmationScreen/BackupSeedPhraseConfirmationScreen';
 export { default as BackupManualScreen } from './BackupManualScreen/BackupManualScreen';
 export { default as BackupRestoreExplanationScreen } from './BackupRestoreExplanationScreen/BackupRestoreExplanationScreen';
